Remove bool dependency for DASH charging.
We're already checking for the file path to determine if dash charging
is available. There's no reason for added check. Especially a check that
really didn't make sense in the first place. At no point are we going to
turn this off on a device that has support for it.

Change-Id: I881ff0ba3f88da4c85858e6667713e701637348b